修正中心化合约

+++ centergame/centergame.hpp
@@ -34,7 +34,7 @@ namespace godapp {
         ACTION init();
         ACTION setglobal(uint64_t key, uint64_t value);
         ACTION transfer(name from, name to, asset quantity, string memo);
-        ACTION reveal(uint8_t game_id, const std::string& message, const std::vector<uint8_t>& bet_ids,
+        ACTION reveal(uint64_t game_id, const std::string& message, const std::vector<uint64_t>& bet_ids,
             const std::vector<asset>& prize_amounts);
         ACTION payment(uint64_t id, name player, name referer, const std::string& message, asset bet, asset payout);
